+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.1
+- __BREAKING CHANGE__:Renames `DirectoryCopyFilter` to `DirectoryFilter` and `FileCopyFilter` to `FileFilter`
+- Creates new method, `filterDirectory()`, for applying the same great filtering logic to in-place-edits (no need to create tmp directories etc.)
+- Creates new method, `deleteEmptyDirectories()`, for deleting all of the empty subdirectories within a directory
+  
 # 1.2.2
 - Won't terminate `sharedStdIn` if `ProcessManager` is not used.